Make things smoother and cleaner
