+++ tests/main.c
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/15 13:51:41 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/10/03 14:29:18 by sbrucker         ###   ########.fr       */
+/*   Updated: 2018/10/08 14:34:06 by sbrucker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,12 +38,12 @@ int	main(int argc, char **argv, char **envp)
 	/*backup_stdout = dup(STDOUT_FILENO);
 	backup_stderr = dup(STDERR_FILENO);*/
 
-	//start = clock();
-	//lexer_tests();
-	//ast_tests();
-	//exec_tests(&envp);
-	//syntax_highlighting_tests(envp);
-	//script_tests(envp);
+	start = clock();
+	lexer_tests();
+	ast_tests();
+	exec_tests(&envp);
+	syntax_highlighting_tests(envp);
+	script_tests(envp);
 	builtin_test_tests(envp);
 
 	/*
